Refactor ServerEnvironment
This is a big refactor, consisting of two parts:
- Switched from a class with static final field to an enum.
- Removed the fork hierarchy.

The switch to an enum was done since there shouldn't be an expectation
of plugin developers adding their own environments. If there's an
environment they need to check, it should be PRed into the main repo.

The fork hierarchy was removed since it was also entirely unnecessary in
the first place. Instead of checking if the current environment is
"at least" something, you simply check if it is that thing. For example,
isPaper() will return true for Folia and Purpur.